requirements.txt
numpy
pandas
scikit-learn
tensorflow
tensorflow.keras
flask
nltk
opencv-python